REF: move populate_backend to superscore.core
